Remove external output API
The eom module used by the external output API has been removed in Tizen 10.0.
https://github.com/flutter-tizen/embedder/issues/84
This will cause dependency issues when launching an app that is already in service on Tizen 10.0.

External output is only used by EvasGL renderer type that uses Elementary window, and stakeholders who requested this feature are no longer interested in this.
Therefore, this feature will be removed and can be provided again based on ecore wl2 API when there is a demand in the future.